Excretes Almost Finished; Will Be Finalized Shortly Tommorow
We continued to go into further detail in ensuing the website is up to the task of fulfilling the needs it is expected, mostly evolving added in CSS in the form of JavaScript to add in the buttons to switch between modes. Very little is left, so it should be done quickly by the beginning of tomorrow.